move TestCustomisation into its own file
